refactor(Components): replace .reaction and .learnMore divs to Components
