+++ src/Components/Elements/RemoveBtn/index.js
@@ -1,0 +1,3 @@
+import RemoveBtn from './RemoveBtn';
+
+export default RemoveBtn;
