+++ src/pages/PostPage.jsx
@@ -1,15 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { reactionRemoved, reactionAdded } from '../Components/Elements/Reactions/reactionsSlice';
+
+import Reactions from '../Components/Elements/Reactions';
 
 import back from '../assets/icons/back.svg';
 import cat from '../assets/img/cat.jpg';
-import like from '../assets/icons/like.svg';
-import likeActive from '../assets/icons/likeActive.svg';
-import dislike from '../assets/icons/dislike.svg';
-import dislikeActive from '../assets/icons/dislikeActive.svg';
 
 const SinglePostWrapper = styled.section`
   padding: 80px 0 64px;
@@ -97,106 +93,6 @@ function SinglePost() {
     id,
   } = location.state;
 
-  const dispatch = useDispatch();
-  const postsWithReactions = useSelector((state) => state.reactions.postsWithReactions);
-
-  const [reactions, setReactions] = useState({
-    countOfLike: Math.round(Math.random() * 50),
-    countOfDislike: Math.round(Math.random() * 50),
-    isLikeActive: false,
-    isDislikeActive: false,
-  });
-
-  const {
-    countOfLike,
-    countOfDislike,
-    isLikeActive,
-    isDislikeActive,
-  } = reactions;
-
-  useEffect(() => {
-    const reaction = postsWithReactions.find((react) => react.id === id);
-    if (reaction) {
-      setReactions({
-        ...reactions,
-        countOfLike: reaction.countOfLike,
-        countOfDislike: reaction.countOfDislike,
-        isLikeActive: reaction.likeStatus,
-        isDislikeActive: reaction.disLikeStatus,
-      });
-    }
-  }, [id, postsWithReactions]);
-
-  const toggleLike = () => {
-    const newLikeStatus = !isLikeActive;
-    const newDislikeStatus = false;
-    let newCountOfLike = countOfLike;
-    let newCountOfDislike = countOfDislike;
-
-    if (!isLikeActive && newLikeStatus) {
-      newCountOfLike += 1;
-      if (isDislikeActive) {
-        newCountOfDislike -= 1;
-      }
-    } else if (isLikeActive && !newLikeStatus) {
-      newCountOfLike -= 1;
-    }
-
-    setReactions({
-      ...reactions,
-      countOfLike: newCountOfLike,
-      countOfDislike: newCountOfDislike,
-      isLikeActive: newLikeStatus,
-      isDislikeActive: newDislikeStatus,
-    });
-
-    dispatch(reactionRemoved(id));
-    dispatch(
-      reactionAdded({
-        id,
-        likeStatus: newLikeStatus,
-        disLikeStatus: newDislikeStatus,
-        countOfLike: newCountOfLike,
-        countOfDislike: newCountOfDislike,
-      }),
-    );
-  };
-
-  const toggleDislike = () => {
-    const newDislikeStatus = !isDislikeActive;
-    const newLikeStatus = false;
-    let newCountOfDislike = countOfDislike;
-    let newCountOfLike = countOfLike;
-
-    if (!isDislikeActive && newDislikeStatus) {
-      newCountOfDislike += 1;
-      if (isLikeActive) {
-        newCountOfLike -= 1;
-      }
-    } else if (isDislikeActive && !newDislikeStatus) {
-      newCountOfDislike -= 1;
-    }
-
-    setReactions({
-      ...reactions,
-      countOfLike: newCountOfLike,
-      countOfDislike: newCountOfDislike,
-      isLikeActive: newLikeStatus,
-      isDislikeActive: newDislikeStatus,
-    });
-
-    dispatch(reactionRemoved(id));
-    dispatch(
-      reactionAdded({
-        id,
-        likeStatus: newLikeStatus,
-        disLikeStatus: newDislikeStatus,
-        countOfLike: newCountOfLike,
-        countOfDislike: newCountOfDislike,
-      }),
-    );
-  };
-
   return (
     <SinglePostWrapper>
       <header>
@@ -204,28 +100,7 @@ function SinglePost() {
           <img src={back} alt="back to main page" />
           <div>Вернуться к статьям</div>
         </NavLink>
-        <div className="reactions">
-          <div className="reaction__count">
-            <button
-              type="button"
-              aria-label="add like"
-              onClick={() => toggleLike()}
-            >
-              <img src={isLikeActive ? likeActive : like} alt="like" />
-            </button>
-            <p>{countOfLike}</p>
-          </div>
-          <div className="reaction__count">
-            <button
-              type="button"
-              aria-label="add dislike"
-              onClick={() => toggleDislike()}
-            >
-              <img src={isDislikeActive ? dislikeActive : dislike} alt="dislike" />
-            </button>
-            <div>{countOfDislike}</div>
-          </div>
-        </div>
+        <Reactions {...{ id }} />
       </header>
       <h1>{title}</h1>
       <div className="picture">
